Update style.css
Here I fixed a weird bug, the CSS scale property of the Icons was simply not working on google, as if it does not exist. Everything worked fine on Mozila though.
So here's my simple solution: I simply changed the CSS scale property to the CSS transform: scale() property. Here are my change:
dist/style.css
.feature__icon {
scale: 50%; 
} 
to 
.feature__icon {
tranform: scale(50%);
}

And that's all. The icons seems to work fine now :)