Update style.css
Here I fixed a bug on the articles section, taking on account that all the articles are kept inside modals, the bug was on the author's name text. It was my fault but not all my fault. Again, the bug was happening only on google, everything was very fine on Mozilla, but on google part of the text (the author's name) was being hidden out of the modal's. So the fix was very simple and partially my fault. 

The CSS width property on the authors name was like this: 
.modal__artigo--autor {
width:186%;
}
Now, follows the fix:
.modal__artigo--autor {
width:100%
}

It was simple again!! Thank God! :)